Refactored the code a bit. Now the interal pipe is separate from dsbs.cpp.
